Refactor a bit and clean stuff up
